set up MeasureData skeleton (#39)
# Summary

Set up the MeasureData skeleton/demo version. It turns out that Firestore cannot store classes *at all*, so we will have to do it the tedious way of writing our own class to type and type to class conversions of Measures.

# Test Plan

Run unit tests.

-----
Please consider the impact of your changes on the other developers.