let users trash documents without deleting (#40)
# Summary

I added the ability for users to trash documents before the documents are deleted. This way the user can manage the documents their trash without immediately deleting their work.

Only authors should have the capability to delete their documents. Currently, it's not enforced on the backend, but I plan to add that later.  

**Note**: Only the author has access to their trashed documents.

I also added some automated cleanup in the unit tests, so that the database doesn't balloon out of control, even without any users.

# Test Plan

I added a unit test for testing document trashing.

-----
Please consider the impact of your changes on the other developers.